Dont use sudo while creating milestones.
During migration of some bugzilla tickets i noticed random  403
forbidden errors during milestone creation. Just as during issue
creation, the sudo flag is set for creating a milestone too, means it
attempts to create the milestone with the user the current ticket is
processed (and which was missing some access rights in my case).

As we are using an administrative account for migration anyways its not
requird to set sudo value in the headers and i think it is perfectly
fine to use the project's administrative account to create them.